+++ back-end/src/OpenAI/query.js
@@ -32,6 +32,22 @@ function trimPrompt(prompt, maxTokens) {
 }
 
 /**
+ * @function constructCompletionPrompt
+ * @param {string} context - The context for the query
+ * @param {string} query - The query to be completed
+ * @param {string} language - The language for the response
+ * @returns {string} - The completed prompt
+ */
+function constructCompletionPrompt(context, query, language) {
+	const basePrompt = `// Language: javascript\n\n// Request:\n// Create a for loop from 12724 to 889005, and print the current number\n\n// Response:\nfor (let i = 12724; i < 889005; i++) {\n    console.log(i);\n}\n\n// Language: python\n\n// Request:\ndef fibo\n\n// Response:\nnacci(n):\n    if (n == 0):\n        return 0;\n\n    if (n <= 2):\n        return 1;\n \n    return fibonacci(n - 1) + fibonacci(n - 2);\n// Language: javascript\n\n// Request:\nfunction binarySea\n\n// Response:\nrch(array, target) {\n    let low = 0;\n    let high = target.length;\n    \n    while (low <= high) {\n        const middle = Math.floor(low + (high - low) / 2);\n        const middleValue = array[middle]'\n        \n        if (middleValue === target)\n            return middle;\n            \n        if (middleValue > target)\n            low = middle + 1;\n        else\n            high = middle - 1;\n    }\n    \n    return -1;\n}\n\n`;
+
+	return (
+		basePrompt +
+		`// Language: ${language}\n\n// Request:\n${context}\n${query}\n\n// Response:\n`
+	);
+}
+
+/**
  * Query openAI
  * @param {object} body
  * @param {string} body.prompt
@@ -40,7 +56,12 @@ function trimPrompt(prompt, maxTokens) {
  * @param {string} body.stop
  */
 async function queryOpenAI(body) {
-	const nTokens = encode(body.prompt).length;
+	const prompt = constructCompletionPrompt(
+		body.context,
+		body.prompt,
+		body.language
+	);
+	const nTokens = encode(prompt).length;
 	// Increment the user's token count and usage count
 
 	await web.incrementUserData(body.user, {
@@ -50,7 +71,7 @@ async function queryOpenAI(body) {
 	});
 
 	return await openai.createCompletion('code-cushman-001', {
-		prompt: body.prompt,
+		prompt: prompt,
 		temperature: 0,
 		max_tokens: clamp(MAX_TOKENS - nTokens, 1, body.maxTokens),
 		stop: body.stop,