+++ src/languages.js
@@ -1,0 +1,123 @@
+/*
+	author....: nekumelon
+	License...: MIT (Check LICENSE)
+*/
+
+const languages = {
+	javascript: {
+		comment: '//'
+	},
+	jsx: {
+		comment: '//'
+	},
+	html: {
+		comment: '<!--'
+	},
+	css: {
+		comment: '/*'
+	},
+	scss: {
+		comment: '/*'
+	},
+	less: {
+		comment: '/*'
+	},
+	graphql: {
+		comment: '#'
+	},
+	json: {
+		comment: '//'
+	},
+	yaml: {
+		comment: '#'
+	},
+	yml: {
+		comment: '#'
+	},
+	vue: {
+		comment: '<!--'
+	},
+	angular: {
+		comment: '<!--'
+	},
+	typescript: {
+		comment: '//'
+	},
+	flow: {
+		comment: '//'
+	},
+	md: {
+		comment: '<!--'
+	},
+	mdx: {
+		comment: '<!--'
+	},
+	gfm: {
+		comment: '<!--'
+	},
+	go: {
+		comment: '//'
+	},
+	rust: {
+		comment: '//'
+	},
+	c: {
+		comment: '//'
+	},
+	cpp: {
+		comment: '//'
+	},
+	csharp: {
+		comment: '//'
+	},
+	java: {
+		comment: '//'
+	},
+	kotlin: {
+		comment: '//'
+	},
+	php: {
+		comment: '//'
+	},
+	python: {
+		comment: '#'
+	},
+	ruby: {
+		comment: '#'
+	},
+	swift: {
+		comment: '//'
+	},
+	objectivec: {
+		comment: '//'
+	},
+	objectivecpp: {
+		comment: '//'
+	},
+	lua: {
+		comment: '--'
+	},
+	haskell: {
+		comment: '--'
+	},
+	clojure: {
+		comment: ';'
+	},
+	elixir: {
+		comment: ';'
+	},
+	erlang: {
+		comment: '%'
+	},
+	elm: {
+		comment: '--'
+	},
+	fsharp: {
+		comment: '//'
+	},
+	default: {
+		comment: '//'
+	}
+};
+
+module.exports = { languages };