+++ src/query/query.js
@@ -8,24 +8,22 @@ const vscode = require('vscode');
 
 const config = vscode.workspace.getConfiguration('codexr');
 
-// function constructInsertPrompt(prefix, line, suffix, language) {
-//     return `// Language: ${language}\n${prefix}\n${line}[insert]\n${suffix}`;
-// }
-
-function constructCompletionPrompt(context, query, language) {
-	const basePrompt = `// Language: javascript\n\n// Request:\n// Create a for loop from 12724 to 889005, and print the current number\n\n// Response:\nfor (let i = 12724; i < 889005; i++) {\n    console.log(i);\n}\n\n// Language: python\n\n// Request:\ndef fibo\n\n// Response:\nnacci(n):\n    if (n == 0):\n        return 0;\n\n    if (n <= 2):\n        return 1;\n \n    return fibonacci(n - 1) + fibonacci(n - 2);\n// Language: javascript\n\n// Request:\nfunction binarySea\n\n// Response:\nrch(array, target) {\n    let low = 0;\n    let high = target.length;\n    \n    while (low <= high) {\n        const middle = Math.floor(low + (high - low) / 2);\n        const middleValue = array[middle]'\n        \n        if (middleValue === target)\n            return middle;\n            \n        if (middleValue > target)\n            low = middle + 1;\n        else\n            high = middle - 1;\n    }\n    \n    return -1;\n}\n\n`;
-
-	return (
-		basePrompt +
-		`// Language: ${language}\n\n// Request:\n${context}\n${query}\n\n// Response:\n`
-	);
-}
-
+/**
+ * @async
+ * @function queryOpenAI
+ * @param {string} context - The context of the query
+ * @param {string} query - The query
+ * @param {string} user - The user
+ * @param {string} language - The language
+ * @returns {Promise} - The response from the server
+ */
 async function queryOpenAI(context, query, user, language) {
 	const maxTokens = config.get('max_tokens');
 
 	return await axios.post('https://codexr.herokuapp.com/query', {
-		prompt: constructCompletionPrompt(context, query, language),
+		prompt: query,
+		language: language,
+		context: context,
 		stop: ['\n\n\n', '// Language:', '// Request:', '// Response:'],
 		user: user,
 		maxTokens: maxTokens